mjcs/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-admin==2.0.1
django-ckeditor==6.5.1
django-colorfield==0.7.2
django-environ==0.9.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-js-asset==2.0.0
django-six==1.0.4
gunicorn==20.1.0
Pillow==9.2.0
pytz==2022.4
screen==1.0.1
sqlparse==0.4.3
xlwt==1.3.0
